refactor: split go mod into 2 pipes
