+++ internal/pipe/gomod/doc.go
@@ -1,0 +1,3 @@
+// Package gomod provides go modules utilities, such as template variables and the ability to proxy the module from
+// proxy.golang.org.
+package gomod
